+++ src/components/AppBottomSheet.tsx
@@ -1,21 +1,22 @@
 import AppBottomSheetBackdrop from './AppBottomSheetBackdrop';
 import BottomSheet, { BottomSheetScrollView, type BottomSheetProps } from '@gorhom/bottom-sheet';
 import React, { useEffect, useMemo, useRef, useState, type ReactNode } from 'react';
-import { Dimensions, View } from 'react-native';
+import { Dimensions, Platform, View } from 'react-native';
 import { type StyleProps } from 'react-native-reanimated';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Fader } from 'react-native-ui-lib';
 import tw from 'twrnc';
 
 const HANDLE_HEIGHT = 8;
-const MIN_MARGIN_BOTTOM = 16;
+const MIN_PADDING_BOTTOM = 20;
+const MIN_BACKDROP_HEIGHT = 64;
 
 export type AppBottomSheetProps = Omit<BottomSheetProps, 'snapPoints'> & {
   children?: ReactNode;
   contentContainerStyle?: StyleProps;
 };
 
-const { height: windowHeight } = Dimensions.get('window');
+const { height: screenHeight } = Dimensions.get('screen');
 
 const AppBottomSheet = ({
   children,
@@ -26,19 +27,18 @@ const AppBottomSheet = ({
   const insets = useSafeAreaInsets();
   const bottomSheetRef = useRef<BottomSheet>(null);
   const [contentHeight, setContentHeight] = useState(0);
-  const snapPoints = useMemo(
-    () => [
-      Math.min(
-        Math.max(contentHeight + HANDLE_HEIGHT + MIN_MARGIN_BOTTOM, 256),
-        windowHeight - MIN_MARGIN_BOTTOM * 2 - insets.bottom - insets.top,
-      ),
-    ],
-    [contentHeight, insets],
-  );
 
+  const visibleHeight = useMemo(
+    () => Math.max(contentHeight + HANDLE_HEIGHT, 256),
+    [contentHeight],
+  );
+  const maxHeight = useMemo(
+    () => screenHeight - MIN_BACKDROP_HEIGHT - insets.bottom - insets.top,
+    [screenHeight, insets],
+  );
   const isBouncing = useMemo(() => {
-    return contentHeight > windowHeight - MIN_MARGIN_BOTTOM * 2 - insets.bottom - insets.top;
-  }, [contentHeight, insets]);
+    return visibleHeight > maxHeight;
+  }, [visibleHeight, maxHeight]);
 
   useEffect(() => {
     if (contentHeight) {
@@ -62,14 +62,18 @@ const AppBottomSheet = ({
         />
       )}
       backgroundStyle={tw`bg-white dark:bg-zinc-900`}
-      bottomInset={insets.bottom || MIN_MARGIN_BOTTOM}
+      bottomInset={4}
       containerStyle={tw`z-10`}
       detached={true}
       handleIndicatorStyle={tw`bg-gray-500 rounded-full`}
       handleStyle={tw`bg-transparent absolute right-0 left-0`}
       {...props}
-      snapPoints={snapPoints}
-      style={[tw`mx-4 rounded-[2rem] overflow-hidden relative`, style]}>
+      snapPoints={[Math.min(visibleHeight, maxHeight)]}
+      style={[
+        tw`mx-1 overflow-hidden relative`,
+        Platform.OS === 'android' ? tw`rounded-[2rem]` : tw`rounded-[3rem]`,
+        style,
+      ]}>
       <View style={tw`absolute top-0 left-0 right-0 z-10`}>
         <Fader
           position={Fader.position.TOP}
@@ -80,7 +84,11 @@ const AppBottomSheet = ({
       {children && (
         <BottomSheetScrollView
           bounces={isBouncing}
-          contentContainerStyle={[tw`pt-2`, isBouncing && tw`pb-6`, contentContainerStyle]}
+          contentContainerStyle={[
+            tw`pt-2`,
+            { paddingBottom: insets.bottom || MIN_PADDING_BOTTOM },
+            contentContainerStyle,
+          ]}
           onContentSizeChange={(_width, height) => setContentHeight(height)}>
           {children}
         </BottomSheetScrollView>